Minor change to ViewVenue method.

+++ BuildYourEvent/src/BuildYourEvent/Controllers/HomeController.cs
@@ -253,7 +253,7 @@ namespace BuildYourEvent.Controllers
             return RedirectToAction("Index");
         }
 
-        public IActionResult ViewVenue(int id)
+        public IActionResult ViewVenue(short id)
         {
             dynamic model = new ExpandoObject();
             var venue = (from v in _context.Venues where v.id == id select v).FirstOrDefault();
